refactor(ui): Extract handling of default values in the form
